feat: added exit button + moved the "what would you like to make text"
