+++ gui/src/pages/creator/index.tsx
@@ -5,69 +5,72 @@ import { PlanEditor } from "./planEditor"
 import { InputBox } from "./inputBox"
 import "./ui/index.css";
 import { useMessaging } from "@/util/messagingContext"
+import { PearIcon } from "./ui/pearIcon"
+import { Button } from "./ui/button"
+import { LogOut } from "lucide-react"
 
 
 
 // TODO: refactor - works for now and I'm too scared to touch it - James
 // Animation info stored in window to survive component remounts
 if (typeof window !== 'undefined') {
-  window.__creatorOverlayAnimation = window.__creatorOverlayAnimation || {
-    direction: null,
-    timestamp: 0
-  };
+	window.__creatorOverlayAnimation = window.__creatorOverlayAnimation || {
+		direction: null,
+		timestamp: 0
+	};
 }
 
 const getAnimationDirection = () => {
-  if (typeof window === 'undefined') return null;
-  // If it's been more than 1 second since the last animation update,
-  // assume we're in a stable state based on the last known direction
-  const now = Date.now();
-  const timeSinceUpdate = now - window.__creatorOverlayAnimation.timestamp;
-  
-  if (timeSinceUpdate > 1000) {
-    // If last direction was "up", we should be hidden (-100%)
-    // If last direction was "down", we should be visible (0)
-    // If no direction yet, default null
-    return window.__creatorOverlayAnimation.direction;
-  }
-  
-  // We're in an active animation, return the current direction
-  return window.__creatorOverlayAnimation.direction;
+	if (typeof window === 'undefined') return null;
+	// If it's been more than 1 second since the last animation update,
+	// assume we're in a stable state based on the last known direction
+	const now = Date.now();
+	const timeSinceUpdate = now - window.__creatorOverlayAnimation.timestamp;
+
+	if (timeSinceUpdate > 1000) {
+		// If last direction was "up", we should be hidden (-100%)
+		// If last direction was "down", we should be visible (0)
+		// If no direction yet, default null
+		return window.__creatorOverlayAnimation.direction;
+	}
+
+	// We're in an active animation, return the current direction
+	return window.__creatorOverlayAnimation.direction;
 };
 
 const setAnimationDirection = (direction) => {
-  if (typeof window === 'undefined') return;
-  window.__creatorOverlayAnimation = {
-    direction,
-    timestamp: Date.now()
-  };
+	if (typeof window === 'undefined') return;
+	window.__creatorOverlayAnimation = {
+		direction,
+		timestamp: Date.now()
+	};
 };
 
 type ExtensionMessage =
-  | { data: { plan: string } }
-  | { messageType: "planCreationSuccess" }
-  | { messageType: "pearAiCloseCreatorInterface" }
-  | { messageType: "pearAiHideCreatorLoadingOverlay" }
-  | { messageType: "newCreatorModeTask"; text: string }
-  | { messageType: "creatorModePlannedTaskSubmit"; text: string }
-  | { messageType: "overlayAnimation"; data: { direction: "up" | "down" } };
+	| { data: { plan: string } }
+	| { messageType: "planCreationSuccess" }
+	| { messageType: "pearAiCloseCreatorInterface" }
+	| { messageType: "pearAiHideCreatorLoadingOverlay" }
+	| { messageType: "newCreatorModeTask"; text: string }
+	| { messageType: "creatorModePlannedTaskSubmit"; text: string }
+	| { messageType: "overlayAnimation"; data: { direction: "up" | "down" } };
 
 /**
  * CreatorOverlay component provides a full-screen overlay with an auto-focusing input field
  * for capturing user commands or queries.
  */
 export const CreatorOverlay = () => {
-  
+
 	const [initialMessage, setInitialMessage] = useState("")
 	const [newProjectPlan, setNewProjectPlan] = useState("")
 	const [currentState, setCurrentState] = useState<"IDEATION" | "GENERATING_PLAN" | "GENERATED_PLAN">("IDEATION")
-	
+
 	// Keep animation state in a ref to prevent render cycles
 	const animationRef = useRef(getAnimationDirection());
-	
+
 	// Force a rerender when animation changes 
 	const [, forceUpdate] = useState({});
-	
+
 	const textareaRef = useRef<HTMLTextAreaElement | null>(null)
 	const isCapturingRef = useRef(false)
 	const { sendMessage, typedRegister, registerListener } = useMessaging();
@@ -163,17 +166,17 @@ export const CreatorOverlay = () => {
 	useEffect(() => {
 		const handleAnimation = (msg) => {
 			if (!msg.data?.direction) return;
-			
+
 			const newDirection = msg.data.direction;
 			// Store direction in window for persistence
 			setAnimationDirection(newDirection);
-			
+
 			// Update our ref
 			animationRef.current = newDirection;
-			
+
 			// Force a rerender
 			forceUpdate({});
-			
+
 			// Direct DOM manipulation as fallback
 			try {
 				// Find the main animation container in DOM
@@ -188,7 +191,7 @@ export const CreatorOverlay = () => {
 
 		// Register handler
 		const unregister = registerListener("overlayAnimation", handleAnimation);
-		
+
 		return () => {
 			if (typeof unregister === 'function') {
 				unregister();
@@ -220,15 +223,25 @@ export const CreatorOverlay = () => {
 
 	return (
 		<div className="w-full h-full" data-animation-direction={getAnimationDirection()}>
-			<div 
-				onClick={close} 
+			<div
+				onClick={close}
 				style={getTransformStyle()}
-				className="all-initial fixed inset-0 flex items-center justify-center bg-transparent font[var(--vscode-font-family)] animate"
+				className="all-initial fixed inset-0 flex items-center justify-center bg-transparent font[var(--vscode-font-family)] animate flex-col"
 			>
-				<div 
-					onClick={(e) => e.stopPropagation()} 
-					className="justify-center align-middle m-auto w-full max-w-3xl"
+				<div
+					onClick={(e) => e.stopPropagation()}
+					className="justify-center align-middle m-auto w-full max-w-3xl gap-4 flex flex-col"
 				>
+					{
+						currentState === "IDEATION" && (
+							<div className="flex justify-center align-middle text-black w-full gap-2 text-lg">
+								<PearIcon className="my-auto size-8" />
+								<div className="my-auto">
+									What would you like to make?
+								</div>
+							</div>
+						)
+					}
 					<RGBWrapper className="px-4 my-auto w-full">
 						{/* Stage 1: get the input from the user about what to make */}
 						<InputBox
@@ -255,6 +268,12 @@ export const CreatorOverlay = () => {
 						)}
 					</RGBWrapper>
 				</div>
+				<div >
+				</div>
+				<Button variant="secondary" size="sm" className="mb-8 cursor-pointer">
+					<LogOut className="size-4" />
+					Exit Creator
+				</Button>
 			</div>
 		</div>
 	)
