+++ gui/src/pages/creator/ui/pearIcon.tsx
